requirements.txt
fastapi==0.68.1
pydantic==1.8.2
pymongo==3.12.0
python-dotenv==0.19.1
passlib==1.7.4
jose==1.7.0
pandas==1.3.4
starlette==0.14.2
